% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{accident2014}
\alias{accident2014}
\title{Sample of car accident location in the UK during year 2014.}
\format{The dataset has 500 instances described by 2 variables (coordinates).}
\source{
\url{https://data.gov.uk/}
}
\usage{
accident2014
}
\description{
Longitude and latitude of 500 car accident during year 2014 (source: data.gov.uk).
}
