% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Alcohol dataset}
\format{
The dataset has 27 instances described by 4 variables.
The variables are the average amount of alcohol of different types consumed per year per inhabitant.
}
\source{
\url{https://www.who.int/}
}
\usage{
alcohol
}
\description{
This dataset has been extracted from the WHO database and depicts alcohol consumption habits in 27 European countries (in 2010).
}
