% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{beetles}
\alias{beetles}
\title{Flea beetles dataset}
\format{
The dataset has 74 instances described by 3 variables.
The variables are as follows:
\describe{
\item{\code{Width}}{The maximal width of aedeagus in the forpart (in microns).}
\item{\code{Angle}}{The front angle of the aedeagus (1 unit = 7.5 degrees).}
\item{\code{Species}}{Species of flea beetle from the genus \emph{Chaetocnema}.}
}
}
\source{
Lubischew, A.A. (1962) On the use of discriminant functions in taxonomy. Biometrics, 18, 455-477.
}
\usage{
beetles
}
\description{
Data were collected on the genus of flea beetle \emph{Chaetocnema}, which contains three species:
\emph{concinna}, \emph{heikertingeri}, and \emph{heptapotamica}.
Measurements were made on the width and angle of the aedeagus of each beetle.
The goal of the original study was to form a classification rule to distinguish the three species.
}
