% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{boosting-class}
\alias{boosting-class}
\title{Boosting methods model}
\description{
This class contains the ensemble of models obtained by a boosting or bagging method
(\code{\link{ADABOOST}}, \code{\link{BAGGING}}).
}
\details{
Objects of this class are plain lists with the following components:
\describe{
  \item{\code{models}}{List of models.}
  \item{\code{x}}{The learning set.}
  \item{\code{y}}{The target values.}
  \item{\code{nsamples}}{The number of models that were asked for. Boosting keeps fewer
    when it runs out of models better than chance; \code{\link{print.boosting}} says so.}
}
}
\seealso{
\code{\link{ADABOOST}}, \code{\link{BAGGING}}, \code{\link{predict.boosting}}
}
