% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{britpop}
\alias{britpop}
\title{Population and location of 18 major british cities.}
\format{The dataset has 18 instances described by 3 variables.}
\usage{
britpop
}
\description{
Longitude and latitude and population of 18 major cities in the Great Britain.
}
