% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{capitals}
\alias{capitals}
\title{Capitals dataset}
\format{
A character \code{vector} of 30 sentences, lowercase and free of punctuation.
}
\usage{
capitals
}
\description{
A small corpus of English sentences about France, Germany and their capitals.
It is deliberately tiny, so that the examples of the text mining functions run in a moment;
a real corpus is loaded with \code{\link{loadtext}}.
}
\seealso{
\code{\link{getvocab}}, \code{\link{vectorize.docs}}, \code{\link{vectorize.words}},
\code{\link{loadtext}}
}
\author{
Alexandre Blansché \email{alexandre.blansche@univ-lorraine.fr}
}
