% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cartdepth}
\alias{cartdepth}
\title{Depth}
\usage{
cartdepth(model)
}
\arguments{
\item{model}{The decision tree.}
}
\value{
The depth.
}
\description{
Return the depth of a decision tree.
}
\examples{
require (datasets)
data (iris)
model = CART (iris [, -5], iris [, 5])
cartdepth (model)
}
\seealso{
\code{\link{CART}}, \code{\link{cartinfo}}, \code{\link{cartleafs}}, \code{\link{cartnodes}}, \code{\link{cartplot}}
}
