% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cartinfo}
\alias{cartinfo}
\title{CART information}
\usage{
cartinfo(model)
}
\arguments{
\item{model}{The decision tree.}
}
\value{
Various information organized into a vector.
}
\description{
Return various information on a CART model.
}
\examples{
require (datasets)
data (iris)
model = CART (iris [, -5], iris [, 5])
cartinfo (model)
}
\seealso{
\code{\link{CART}}, \code{\link{cartdepth}}, \code{\link{cartleafs}}, \code{\link{cartnodes}}, \code{\link{cartplot}}
}
