% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cartleafs}
\alias{cartleafs}
\title{Number of Leafs}
\usage{
cartleafs(model)
}
\arguments{
\item{model}{The decision tree.}
}
\value{
The number of leafs.
}
\description{
Return the number of leafs of a decision tree.
}
\seealso{
\code{\link{CART}}, \code{\link{cartdepth}}, \code{\link{cartinfo}}, \code{\link{cartnodes}}, \code{\link{cartplot}}
require (datasets)
data (iris)
model = CART (iris [, -5], iris [, 5])
cartleafs (model)
}
