% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cartnodes}
\alias{cartnodes}
\title{Number of Nodes}
\usage{
cartnodes(model)
}
\arguments{
\item{model}{The decision tree.}
}
\value{
The number of nodes.
}
\description{
Return the number of nodes of a decision tree.
}
\seealso{
\code{\link{CART}}, \code{\link{cartdepth}}, \code{\link{cartinfo}}, \code{\link{cartleafs}}, \code{\link{cartplot}}
require (datasets)
data (iris)
model = CART (iris [, -5], iris [, 5])
cartnodes (model)
}
