% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cartplot}
\alias{cartplot}
\title{CART Plot}
\usage{
cartplot(model, ...)
}
\arguments{
\item{model}{The decision tree.}

\item{...}{Other parameters.}
}
\description{
Plot a decision tree obtained by CART.
}
\examples{
require (datasets)
data (iris)
model = CART (iris [, -5], iris [, 5])
cartplot (model)
}
\seealso{
\code{\link{CART}}, \code{\link{cartdepth}}, \code{\link{cartinfo}}, \code{\link{cartleafs}}, \code{\link{cartnodes}}
}
