% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cda-class}
\alias{cda-class}
\title{Canonical Disciminant Analysis model}
\description{
This class contains the classification model obtained by the CDA method.
}
\details{
Objects of this class are plain lists with the following components:
\describe{
  \item{\code{proj}}{The projection of the dataset into the canonical base. A \code{data.frame}.}
  \item{\code{transform}}{The transformation matrix between. A \code{matrix}.}
  \item{\code{centers}}{Coordinates of the class centers. A \code{matrix}.}
  \item{\code{within}}{The intra-class covariance matrix. A \code{matrix}.}
  \item{\code{eig}}{One row per canonical axis and four columns: the \code{eigenvalue}
    (of \eqn{V^{-1}B}, i.e. the squared canonical correlation, between 0 and 1), its
    \code{percentage of variance} (share of the trace) and the \code{cumulative} one, and
    the \code{discriminant power} -- the share of the trace of \eqn{W^{-1}B}, which is what
    \code{\link[MASS]{lda}} and most other software call the proportion of trace. A
    \code{matrix}, or a named \code{vector} when there is a single axis.}
  \item{\code{dim}}{The number of dimensions of the canonical base (numeric value).}
  \item{\code{nb.classes}}{The number of clusters (numeric value).}
  \item{\code{train}}{The training set (description). A \code{data.frame}.}
  \item{\code{labels}}{Class labels of the training set. Either a \code{factor} or an integer \code{vector}.}
  \item{\code{model}}{The prediction model.}
}
}
\seealso{
\code{\link{CDA}}, \code{\link{plot.cda}}, \code{\link{predict.cda}}
}
