% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{correlated}
\alias{correlated}
\title{Correlated variables}
\usage{
correlated(d, threshold = 0.8)
}
\arguments{
\item{d}{A data matrix.}

\item{threshold}{The threshold on the (absolute) Pearson coefficient. If NULL, return the most correlated variables.}
}
\value{
The list of correlated variables (as a matrix of column names).
}
\description{
Return the list of correlated variables
}
\examples{
data (iris)
correlated (iris)
}
\seealso{
\code{\link[stats]{cor}}
}
