% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{credit}
\alias{credit}
\title{Credit dataset}
\format{
The dataset has 66 instances described by 11 qualitative variables.
}
\usage{
credit
}
\description{
This is a fake dataset simulating a bank database about loan clients.
}
