% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{data.parabol}
\alias{data.parabol}
\title{Parabol dataset}
\usage{
data.parabol(
  n = c(500, 100),
  xlim = c(-3, 3),
  center = c(0, 4),
  coeff = 0.5,
  sigma = c(0.5, 0.5),
  levels = NULL,
  graph = FALSE,
  seed = NULL
)
}
\arguments{
\item{n}{Number of observations in each class.}

\item{xlim}{Minimum and maximum on the x axis.}

\item{center}{Coordinates of the center of the gaussian distribution.}

\item{coeff}{Coefficient of the parabol.}

\item{sigma}{Variance in each class.}

\item{levels}{Name of each class.}

\item{graph}{Whether the generated dataset is plotted. \code{FALSE} by default, as
everywhere else in the package: a generator has to be callable in a loop or a report
without piling up graphics devices.}

\item{seed}{A specified seed for random number generation.}
}
\value{
A randomly generated dataset.
}
\description{
Generate a random dataset shaped like a parabol and a gaussian distribution
}
\examples{
data.parabol (graph = TRUE)
}
\seealso{
\code{\link{data.diag}}, \code{\link{data.target1}}, \code{\link{data.target2}}, \code{\link{data.twomoons}}, \code{\link{data.xor}}
}
