% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{data.twomoons}
\alias{data.twomoons}
\title{Two moons dataset}
\usage{
data.twomoons(
  r = 1,
  n = 200,
  sigma = 0.1,
  levels = NULL,
  graph = FALSE,
  seed = NULL
)
}
\arguments{
\item{r}{Radius of each class.}

\item{n}{Number of observations in each class.}

\item{sigma}{Variance in each class.}

\item{levels}{Name of each class.}

\item{graph}{Whether the generated dataset is plotted. \code{FALSE} by default, as
everywhere else in the package: a generator has to be callable in a loop or a report
without piling up graphics devices.}

\item{seed}{A specified seed for random number generation.}
}
\value{
A randomly generated dataset.
}
\description{
Generate a random dataset shaped like two moons.
}
\examples{
data.twomoons (graph = TRUE)
}
\seealso{
\code{\link{data.diag}}, \code{\link{data.parabol}}, \code{\link{data.target1}}, \code{\link{data.target2}}, \code{\link{data.xor}}
}
