% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{data1}
\alias{data1}
\title{"data1" dataset}
\format{240 observations described by 4 variables and grouped into 16 classes.}
\usage{
data1
}
\description{
Synthetic dataset.
}
\author{
Alexandre Blansché \email{alexandre.blansche@univ-lorraine.fr}
}
