% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{data2}
\alias{data2}
\title{"data2" dataset}
\format{
500 observations described by 10 variables and grouped into 3 classes.
}
\usage{
data2
}
\description{
Synthetic dataset.
}
\author{
Alexandre Blansché \email{alexandre.blansche@univ-lorraine.fr}
}
