% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{data3}
\alias{data3}
\title{"data3" dataset}
\format{300 observations described by 3 variables and grouped into 3 classes.}
\usage{
data3
}
\description{
Synthetic dataset.
}
\author{
Alexandre Blansché \email{alexandre.blansche@univ-lorraine.fr}
}
