% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{decathlon}
\alias{decathlon}
\title{Decathlon dataset}
\format{
The dataset has 41 instances described by 13 variables.
The variables are as follows:
\describe{
\item{\code{100m}}{In seconds.}
\item{\code{Long.jump}}{In meters.}
\item{\code{Shot.put}}{In meters.}
\item{\code{High.jump}}{In meters.}
\item{\code{400m}}{In seconds.}
\item{\code{110m.h}}{In seconds.}
\item{\code{Discus.throw}}{In meters.}
\item{\code{Pole.vault}}{In meters.}
\item{\code{Javelin.throw}}{In meters.}
\item{\code{1500m}}{In seconds.}
\item{\code{Rank}}{The rank at the competition.}
\item{\code{Points}}{The number of points obtained by the athlete.}
\item{\code{Competition}}{\code{Olympics} or \code{Decastar}.}
}
}
\source{
\url{https://husson.github.io/data.html}
}
\usage{
decathlon
}
\description{
The dataset contains results from two athletics competitions.
The 2004 Olympic Games in Athens and the 2004 Decastar.
}
