% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{eucalyptus}
\alias{eucalyptus}
\title{Eucalyptus dataset}
\format{
The dataset has 1429 instances (eucalyptus trees) with 2 measurements: the height and the circumference.
}
\source{
\url{http://www.cmap.polytechnique.fr/~lepennec/en/teaching/}
}
\usage{
eucalyptus
}
\description{
Measuring the height of a tree is not an easy task. Is it possible to estimate the height as a function of the circumference of the trunk?
}
