% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{evaluation.kappa}
\alias{evaluation.kappa}
\title{Kappa evaluation of classification predictions}
\usage{
evaluation.kappa(predictions, gt, ...)
}
\arguments{
\item{predictions}{The predictions of a classification model (\code{factor} or \code{vector}).}

\item{gt}{The ground truth (\code{factor} or \code{vector}).}

\item{...}{Other parameters.}
}
\value{
The evaluation of the predictions (numeric value).
}
\description{
Evaluation predictions of a classification model according to Cohen's kappa: the proportion
of correct predictions, corrected for the proportion two independent labellings would get
right by chance. Class labels being nominal, the kappa is the unweighted one -- every
mistake counts the same.
}
\examples{
require (datasets)
data (iris)
d = splitdata (iris, 5)
model.nb = NB (d$train.x, d$train.y)
pred.nb = predict (model.nb, d$test.x)
evaluation.kappa (pred.nb, d$test.y)
}
\seealso{
\code{\link{evaluation.accuracy}}, \code{\link{evaluation.fmeasure}}, \code{\link{evaluation.fowlkesmallows}}, \code{\link{evaluation.goodness}}, \code{\link{evaluation.jaccard}}, \code{\link{evaluation.kappa}}, \code{\link{evaluation.precision}},
\code{\link{evaluation.precision}}, \code{\link{evaluation.recall}},
\code{\link{evaluation}}
}
