% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{exportgraphics}
\alias{exportgraphics}
\title{Open a graphics device}
\usage{
exportgraphics(
  file,
  type = tail(strsplit(file, split = "\\\\.")[[1]], 1),
  export = fdm2id.globals$export,
  ...
)
}
\arguments{
\item{file}{A character string giving the name of the file.}

\item{type}{The type of graphics device. Deduced from the file extension by default:
\code{"eps"} and \code{"ps"} go to \code{\link[grDevices]{postscript}}, \code{"jpg"} to
\code{\link[grDevices]{jpeg}}, and any other extension is taken as the name of an R
function (\code{"pdf"}, \code{"png"}, ...).}

\item{export}{If given, explicitly overrides the global export toggle set by \code{\link{toggleexport}} for
this call only. By default, the global toggle is used, so existing code is unaffected.}

\item{...}{Other parameters.}
}
\description{
Starts the graphics device driver
}
\examples{
\dontrun{
data (iris)
exportgraphics ("export.pdf")
plotdata (iris [, -5], iris [, 5])
closegraphics()
# Extensions that don't match an R function name directly are now handled:
exportgraphics ("export.eps")
plotdata (iris [, -5], iris [, 5])
closegraphics()
}
}
\seealso{
\code{\link{closegraphics}}, \code{\link{toggleexport}}, \code{\link[grDevices]{Devices}}
}
