% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{factorial-class}
\alias{factorial-class}
\title{Factorial analysis results}
\description{
This class contains the result of a factorial analysis, as obtained by \code{\link{CA}},
\code{\link{MCA}} or \code{\link{PCA}}.
}
\details{
Objects of this class are the objects returned by the corresponding \pkg{FactoMineR}
functions (\code{\link[FactoMineR]{CA}}, \code{\link[FactoMineR]{MCA}} and
\code{\link[FactoMineR]{PCA}}), with the extra class \code{factorial} prepended so that
\code{\link{plot.factorial}} can provide a uniform plotting interface. The second class
(\code{"ca"}, \code{"mca"} or \code{"pca"}) records which analysis was performed.
}
\seealso{
\code{\link{CA}}, \code{\link{MCA}}, \code{\link{PCA}}, \code{\link{plot.factorial}}
}
