% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{HCA}
\alias{HCA}
\title{Hierarchical Cluster Analysis method}
\usage{
HCA(
  d,
  k = NULL,
  method = c("ward", "single"),
  engine = c("hclust", "agnes"),
  graph = FALSE,
  ...
)
}
\arguments{
\item{d}{The dataset (\code{matrix} or \code{data.frame}).}

\item{k}{The number of cluster. If \code{NULL} (the default), it is set to the largest drop
in aggregation height.}

\item{method}{Character string defining the clustering method.}

\item{engine}{Which implementation builds the hierarchy: \code{\link[stats]{hclust}} (the
default) or \code{\link[cluster]{agnes}}. They give the same hierarchy -- same heights, to
machine precision, and the same cut -- but \code{hclust} is far faster on a large dataset
(0.3 s against 30 s on 3000 observations). Use \code{"agnes"} for the linkages it alone
provides, or to compare the two.}

\item{graph}{A logical indicating whether or not a graphic should be plotted (the
aggregation heights used to choose \code{k}, when \code{k} is not given).}

\item{...}{Other parameters.}
}
\value{
The cluster hierarchy (\code{hca} object).
}
\description{
Run the HCA method for clustering.
}
\examples{
require (datasets)
data (iris)
HCA (iris [, -5], k = 3, method = "ward")
}
\seealso{
\code{\link[stats]{hclust}}, \code{\link[cluster]{agnes}}, \code{\link{treeplot}},
\code{\link{predict.hca}}
}
