% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{kaiser}
\alias{kaiser}
\title{Kaiser rule}
\usage{
kaiser(pca)
}
\arguments{
\item{pca}{The PCA result (object of class \code{factorial-class}).}
}
\description{
Apply the Kaiser rule to determine the appropriate number of PCA axes.
}
\examples{
require (datasets)
data (iris)
pca = PCA (iris, quali.sup = 5)
kaiser (pca)
}
\seealso{
\code{\link{PCA}}, \code{\link{factorial-class}}
}
