% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{keiser}
\alias{keiser}
\title{Keiser rule}
\usage{
keiser(pca)
}
\arguments{
\item{pca}{The PCA result (object of class \code{factorial-class}).}
}
\description{
Apply the keiser rule to determine the appropriate number of PCA axes.
}
\examples{
require (datasets)
data (iris)
pca = PCA (iris, quali.sup = 5)
keiser (pca)
}
\seealso{
\code{\link{PCA}}, \code{\link{factorial-class}}
}
