% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{knn-class}
\alias{knn-class}
\title{K Nearest Neighbours model}
\description{
This class contains the classification model obtained by the k-NN method.
}
\details{
Objects of this class are plain lists with the following components:
\describe{
  \item{\code{train}}{The training set (description). A \code{data.frame}.}
  \item{\code{labels}}{Class labels of the training set. Either a \code{factor} or an integer \code{vector}.}
  \item{\code{k}}{The \code{k} parameter.}
}
}
\seealso{
\code{\link{KNN}}, \code{\link{predict.knn}}
}
