% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{leverageplot}
\alias{leverageplot}
\title{Plot the leverage points of a linear regression model}
\usage{
leverageplot(model, index = NULL, labels = NULL)
}
\arguments{
\item{model}{The model to be plotted.}

\item{index}{The index of the variable used for the x-axis.}

\item{labels}{The labels of the instances.}
}
\description{
Plot the leverage points of a linear regression model.
}
\examples{
require (datasets)
data (trees)
model = LINREG (trees [, -3], trees [, 3])
leverageplot (model)
}
