% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{linsep}
\alias{linsep}
\title{Linsep dataset}
\format{
Class \code{A} contains 50 observations and class \code{B} contains 500 observations.
There are two numeric variables: \code{X} and \code{Y}.
}
\usage{
linsep
}
\description{
Synthetic dataset.
}
\author{
Alexandre Blansché \email{alexandre.blansche@univ-lorraine.fr}
}
