% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{meanshift-class}
\alias{meanshift-class}
\title{MeanShift model}
\description{
This class contains the model obtained by the MEANSHIFT method.
}
\details{
Objects of this class are plain lists with the following components:
\describe{
  \item{\code{cluster}}{A vector of integers indicating the cluster to which each point is allocated.}
  \item{\code{value}}{A vector or matrix containing the location of the classified local maxima in the support.}
  \item{\code{data}}{The leaning set.}
  \item{\code{kernel}}{A string indicating the kernel associated with the kernel density estimate that the mean shift is optimizing over.}
  \item{\code{bandwidth}}{Used in the kernel density estimate for steepest ascent classification.}
  \item{\code{alpha}}{A scalar tuning parameter for normal kernels.}
  \item{\code{iterations}}{The number of iterations to perform mean shift.}
  \item{\code{epsilon}}{A scalar used to determine when to terminate the iteration of an individual query point.}
  \item{\code{epsilonCluster}}{A scalar used to determine the minimum distance between distinct clusters.}
}
}
\seealso{
\code{\link{MEANSHIFT}}
}
