% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{model-class}
\alias{model-class}
\title{Generic classification or regression model}
\description{
This is a wrapper class containing the classification model obtained by any classification or regression method.
}
\details{
Objects of this class are plain lists with the following components:
\describe{
  \item{\code{model}}{The wrapped model.}
  \item{\code{method}}{The name of the method.}
}
}
\seealso{
\code{\link{predict.model}}, \code{\link[stats]{predict}}
}
