% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{model-methods}
\alias{model-methods}
\alias{residuals.model}
\alias{fitted.model}
\alias{rstudent.model}
\title{Residuals and fitted values of a model}
\usage{
\method{residuals}{model}(object, ...)

\method{fitted}{model}(object, ...)

\method{rstudent}{model}(model, ...)
}
\arguments{
\item{object, model}{The model (of class \code{model}).}

\item{...}{Other parameters, passed to the method of the underlying model.}
}
\value{
A numeric vector.
}
\description{
Extract the residuals, the fitted values or the studentized residuals of a model built by one
of the learning methods of the package (\code{\link{LINREG}}, for instance), as they would be
from the underlying model, kept in \code{model$model}.
}
\examples{
require (datasets)
data (trees)
model = LINREG (trees [, -3], trees [, 3])
head (residuals (model))
head (fitted (model))
head (rstudent (model))
}
\seealso{
\code{\link[stats]{residuals}}, \code{\link[stats]{fitted}},
\code{\link[stats]{rstudent}}, \code{\link{resplot}}
}
