% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{movies}
\alias{movies}
\title{Movies dataset}
\format{
A \code{matrix} of 49 films by 55 viewers. Ratings are whole numbers from 1 to 10.
}
\source{
The structure was estimated on the MovieLens 100K dataset
(\url{https://grouplens.org/datasets/movielens/100k/}), whose terms do not permit
redistributing the ratings themselves.
}
\usage{
movies
}
\description{
Simulated ratings of 49 real films by 55 imaginary viewers.
}
