% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{NMF}
\alias{NMF}
\title{Non-negative Matrix Factorization}
\usage{
NMF(x, rank = 2, nstart = 10, seed = NULL, ...)
}
\arguments{
\item{x}{A numeric dataset (data.frame or matrix).}

\item{rank}{Specification of the factorization rank.}

\item{nstart}{How many random sets should be chosen?}

\item{seed}{A specified seed for random number generation.}

\item{...}{Other parameters.}
}
\description{
Return the NMF decomposition.
}
\examples{
\dontrun{
install.packages ("BiocManager")
BiocManager::install ("Biobase")
install.packages ("NMF")
require (datasets)
data (iris)
NMF (iris [, -5])
}
}
\seealso{
\code{\link[NMF]{nmf}}
}
