% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{plot.apriori}
\alias{plot.apriori}
\title{Plot function for apriori-class}
\usage{
\method{plot}{apriori}(
  x,
  method = "scatterplot",
  measure = c("support", "confidence"),
  shading = "lift",
  ...
)
}
\arguments{
\item{x}{The classification model (object of class \code{apriori-class}, created by \code{APRIORI}).}

\item{method}{The type of plot (see \code{\link[arulesViz]{plot}}, e.g. "scatterplot", "graph", "grouped", "paracoord").}

\item{measure, shading}{Parameters passed to \code{\link[arulesViz]{plot}}.}

\item{...}{Other parameters passed to \code{\link[arulesViz]{plot}}.}
}
\description{
Plot the association rules obtained by APRIORI, using arulesViz.
}
\examples{
\dontrun{
require (datasets)
data (iris)
d = discretizeDF (iris,
    default = list (method = "interval", breaks = 3, labels = c ("small", "medium", "large")))
model = APRIORI (d [, -5], d [, 5], supp = .1, conf = .9, prune = TRUE)
plot (model)
plot (model, method = "graph")
}
}
\seealso{
\code{\link{APRIORI}}, \code{\link{apriori-class}}, \code{\link[arulesViz]{plot}}
}
