% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{plot.cda}
\alias{plot.cda}
\title{Plot function for cda-class}
\usage{
\method{plot}{cda}(x, newdata = NULL, axes = 1:2, legendpos = "auto", ...)
}
\arguments{
\item{x}{The classification model (object of class \code{cda-class}).}

\item{newdata}{The test set (\code{matrix} or \code{data.frame}).}

\item{axes}{The canonical axes to be printed (numeric \code{vector}). Ignored when there is
only one, which is what two classes give: the axis is then drawn against the observation
index, and the class means as horizontal lines.}

\item{legendpos}{Position of the legend, as in \code{\link{plotdata}}.}

\item{...}{Other parameters, passed to the underlying plot.}
}
\description{
Plot the learning set (and test set) on the canonical axes obtained by Canonical Discriminant Analysis (function \code{CDA}).
}
\examples{
require (datasets)
data (iris)
model = CDA (iris [, -5], iris [, 5])
plot (model)
}
\seealso{
\code{\link{CDA}}, \code{\link{predict.cda}}, \code{\link{cda-class}}
}
