% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.hca}
\alias{plot.hca}
\title{Plot function for hca-class}
\usage{
\method{plot}{hca}(x, k = x$k, highlight = c("branches", "rectangles"), ...)
}
\arguments{
\item{x}{The hierarchical clustering (created by \code{\link{HCA}}).}

\item{k}{The number of clusters the dendrogram is cut into. Defaults to the cut
\code{\link{HCA}} already made.}

\item{highlight}{How the clusters are highlighted: \code{"branches"} or
\code{"rectangles"} (see \code{\link{treeplot}}).}

\item{...}{Other parameters, passed to \code{\link{treeplot}} (\code{labels},
\code{split}, \code{horiz}, \code{ylab}) or to \code{\link[stats]{plot.dendrogram}}
(for instance \code{axes = FALSE}).}
}
\description{
Draws the dendrogram of a hierarchical clustering, the branches of each cluster in the
colour \code{\link{plotclus}} gives it, so that the dendrogram and the scatter plot of the
same clustering match. The branches above the cut are dashed.
}
\examples{
require (datasets)
data (iris)
hca = HCA (iris [, -5], k = 3, method = "ward")
plot (hca)
plotclus (hca, iris [, -5])
}
\seealso{
\code{\link{HCA}}, \code{\link{treeplot}}, \code{\link{plotclus}}
}
