% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{plotcloud}
\alias{plotcloud}
\title{Plot word cloud}
\usage{
plotcloud(corpus, k = NULL, stopwords = "en", ...)
}
\arguments{
\item{corpus}{The corpus of documents (a vector of characters) or the vocabulary of the documents (result of function \code{getvocab}).}

\item{k}{A categorical variable (vector or factor).}

\item{stopwords}{The language whose stop words are removed (\code{"en"}, ...), or
\code{NULL} to keep them. A list of words of your own goes to \code{excludewords}.}

\item{...}{Other parameters.}
}
\description{
Plot a word cloud based on the word frequencies in the documents.
}
\examples{
data (capitals)
plotcloud (capitals)
vocab = getvocab (capitals, mincount = 1, lang = NULL, stopwords = "en")
plotcloud (vocab)
}
\seealso{
\code{\link{plotzipf}}, \code{\link{getvocab}}, \code{\link[wordcloud]{wordcloud}}
}
