% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{predict.factorial}
\alias{predict.factorial}
\title{Projection of new observations into a factorial space}
\usage{
\method{predict}{factorial}(object, test, ...)
}
\arguments{
\item{object}{The factorial analysis (object of class \code{\link{factorial-class}}).}

\item{test}{The new observations, a \code{data.frame} or \code{matrix} with the same
(active) variables as the data the analysis was fitted on.}

\item{...}{Other parameters.}
}
\value{
The coordinates of the new observations on the factorial axes (a \code{matrix},
one row per observation and one column per axis).
}
\description{
Projects new observations into the factorial space computed by \code{\link{CA}},
\code{\link{MCA}} or \code{\link{PCA}} -- the same operation that is applied to the
observations the analysis was fitted on: centering (and, for \code{\link{PCA}} with
\code{scale.unit = TRUE}, scaling) with the parameters \emph{of the training data}, then
projection on the axes already computed. The axes are not recomputed and the new
observations have no influence on them: they are \emph{supplementary} individuals.
}
\details{
The projection is obtained by handing the new rows back to \pkg{FactoMineR} as supplementary
individuals of the original analysis, so the coordinates are exactly the ones
\code{PCA (rbind (train, test), ind.sup = ...)} would give. The active analysis is refitted
in the process, which is unnoticeable on the sizes this package is meant for.

Supplementary variables (\code{quanti.sup}, \code{quali.sup}) play no part in the axes, so
\code{test} does not have to carry them: any column of the training data that is missing from
\code{test} is filled in (with the training mean, or the first level) purely so that the two
can be stacked.
}
\examples{
require (datasets)
data (iris)
d = splitdata (iris, 5)
pca = PCA (d$train.x)
# The coordinates of unseen observations on the axes of the training analysis
head (predict (pca, d$test.x))
# An observation of the training set projects onto the coordinates the analysis gave it
pca$ind$coord [1, ]
predict (pca, d$train.x [1, ])
}
\seealso{
\code{\link{PCA}}, \code{\link{CA}}, \code{\link{MCA}},
\code{\link{factorial-class}}, \code{\link{predict.cda}}
}
