% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{predict.pam}
\alias{predict.pam}
\title{Predict function for PAM}
\usage{
\method{predict}{pam}(object, newdata, ...)
}
\arguments{
\item{object}{The clustering (created by \code{\link{PAM}}).}

\item{newdata}{A new dataset (a \code{data.frame}), with the same variables as the learning
dataset.}

\item{...}{Other parameters.}
}
\value{
A vector of cluster numbers.
}
\description{
Returns the cluster of the closest medoid, for a new dataset.
}
\examples{
require (datasets)
data (iris)
d = splitdata (iris, 5)
model = PAM (d$train.x, 3)
table (predict (model, d$test.x), d$test.y)
}
\seealso{
\code{\link{PAM}}, \code{\link{predict.kmeans}}
}
