% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{predict.som}
\alias{predict.som}
\title{Predict function for a self-organising map}
\usage{
\method{predict}{som}(object, newdata, ...)
}
\arguments{
\item{object}{The map (created by \code{\link{SOM}}).}

\item{newdata}{A new dataset (a \code{data.frame}), with the same variables as the learning
dataset.}

\item{...}{Other parameters.}
}
\value{
A vector of cluster numbers.
}
\description{
Returns the cluster of the closest unit of the map, for a new dataset.
}
\examples{
\donttest{
require (datasets)
data (iris)
d = splitdata (iris, 5)
model = SOM (d$train.x, 4, 4)
table (predict (model, d$test.x), d$test.y)
}
}
\seealso{
\code{\link{SOM}}, \code{\link{plot.som}}
}
