% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{predict.spectral}
\alias{predict.spectral}
\title{Predict function for Spectral clustering}
\usage{
\method{predict}{spectral}(object, newdata, ...)
}
\arguments{
\item{object}{The clustering model (of class \code{\link{spectral-class}}, created by \code{\link{SPECTRAL}}).}

\item{newdata}{A new dataset (a \code{data.frame}), with same variables as the learning dataset.}

\item{...}{Other parameters.}
}
\description{
Return the closest Spectral clustering cluster for a new dataset. New instances are assigned to
the cluster of their nearest neighbour in the original (training) space, since the spectral
projection cannot be directly extended to unseen data without recomputing the affinity matrix.
}
\examples{
\dontrun{
require (datasets)
data (iris)
d = splitdata (iris, 5)
model = SPECTRAL (d$train.x, k = 3)
predict (model, d$test.x)
}
}
\seealso{
\code{\link{SPECTRAL}}, \code{\link{spectral-class}}
}
