% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{print.boosting}
\alias{print.boosting}
\title{Print an ensemble model}
\usage{
\method{print}{boosting}(x, ...)
}
\arguments{
\item{x}{The model (object of class \code{\link{boosting-class}}, created by
\code{\link{ADABOOST}} or \code{\link{BAGGING}}).}

\item{...}{Other parameters.}
}
\description{
Prints how many models the ensemble holds and what it was fitted on.
}
\examples{
require (datasets)
data (iris)
BAGGING (iris [, -5], iris [, 5], LDA, nsamples = 5)
}
\seealso{
\code{\link{boosting-class}}, \code{\link{ADABOOST}}, \code{\link{BAGGING}}
}
