% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{print.cda}
\alias{print.cda}
\title{Print a canonical discriminant analysis}
\usage{
\method{print}{cda}(x, ...)
}
\arguments{
\item{x}{The model (object of class \code{\link{cda-class}}, created by \code{\link{CDA}}).}

\item{...}{Other parameters.}
}
\description{
Prints the size of the analysis and the variance carried by its axes.
}
\examples{
require (datasets)
data (iris)
CDA (iris [, -5], iris [, 5])
}
\seealso{
\code{\link{cda-class}}, \code{\link{CDA}}, \code{\link{plot.cda}}
}
