% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{print.dataset}
\alias{print.dataset}
\title{Print a training/test split}
\usage{
\method{print}{dataset}(x, ...)
}
\arguments{
\item{x}{The split (object of class \code{\link{dataset-class}}, created by
\code{\link{splitdata}}).}

\item{...}{Other parameters.}
}
\description{
Prints the sizes of the two parts of a split and the target they share.
}
\examples{
require (datasets)
data (iris)
splitdata (iris, 5)
}
\seealso{
\code{\link{dataset-class}}, \code{\link{splitdata}}
}
