% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.dbs}
\alias{print.dbs}
\title{Print a DBSCAN clustering}
\usage{
\method{print}{dbs}(x, ...)
}
\arguments{
\item{x}{The clustering (object of class \code{\link{dbs-class}}, created by
\code{\link{DBSCAN}}).}

\item{...}{Other parameters.}
}
\description{
Prints the number of clusters found, their sizes, the observations left as noise, and the
two parameters used -- instead of dumping the underlying list.
}
\examples{
require (datasets)
data (iris)
DBSCAN (iris [, -5], minpts = 5, eps = 0.65)
}
\seealso{
\code{\link{dbs-class}}, \code{\link{DBSCAN}}
}
