% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.em}
\alias{print.em}
\title{Print an EM clustering}
\usage{
\method{print}{em}(x, ...)
}
\arguments{
\item{x}{The clustering (object of class \code{\link{em-class}}, created by \code{\link{EM}}).}

\item{...}{Other parameters.}
}
\description{
Prints the number of clusters found, their sizes and the log-likelihood reached.
}
\examples{
require (datasets)
data (iris)
EM (iris [, -5], 3)
}
\seealso{
\code{\link{em-class}}, \code{\link{EM}}
}
