% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{print.factorial}
\alias{print.factorial}
\title{Plot function for factorial-class}
\usage{
\method{print}{factorial}(x, ...)
}
\arguments{
\item{x}{The PCA, CA or MCA result (object of class \code{factorial-class}).}

\item{...}{Other parameters.}
}
\description{
Print PCA, CA or MCA.
}
\examples{
require (datasets)
data (iris)
pca = PCA (iris, quali.sup = 5)
print (pca)
}
\seealso{
\code{\link{CA}}, \code{\link{MCA}}, \code{\link{PCA}}, \code{\link[FactoMineR]{print.CA}}, \code{\link[FactoMineR]{print.MCA}}, \code{\link[FactoMineR]{print.PCA}}, \code{\link{factorial-class}}
}
