% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{print.knn}
\alias{print.knn}
\title{Print a \emph{K}-nearest-neighbours model}
\usage{
\method{print}{knn}(x, ...)
}
\arguments{
\item{x}{The model (object of class \code{\link{knn-class}}, created by \code{\link{KNN}}).}

\item{...}{Other parameters.}
}
\description{
Prints the size of the training set the model memorised, and the number of neighbours used.
}
\examples{
require (datasets)
data (iris)
KNN (iris [, -5], iris [, 5])
}
\seealso{
\code{\link{knn-class}}, \code{\link{KNN}}
}
