% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.meanshift}
\alias{print.meanshift}
\title{Print a mean shift clustering}
\usage{
\method{print}{meanshift}(x, ...)
}
\arguments{
\item{x}{The clustering (object of class \code{\link{meanshift-class}}, created by
\code{\link{MEANSHIFT}}).}

\item{...}{Other parameters.}
}
\description{
Prints the number of clusters found, their sizes and the kernel used.
}
\examples{
\donttest{
require (datasets)
data (iris)
MEANSHIFT (iris [, -5])
}
}
\seealso{
\code{\link{meanshift-class}}, \code{\link{MEANSHIFT}}
}
