% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{print.model}
\alias{print.model}
\title{Print a classification or regression model}
\usage{
\method{print}{model}(x, ...)
}
\arguments{
\item{x}{The model to be printed (object of class \code{\link{model-class}}).}

\item{...}{Other parameters.}
}
\description{
Prints a short description of the model -- the method it was obtained with, and the size of
the data it was fitted on -- instead of dumping the underlying list. Use
\code{summary (model)} for the full detail of the wrapped model.
}
\examples{
require (datasets)
data (iris)
NB (iris [, -5], iris [, 5])
}
\seealso{
\code{\link{model-class}}, \code{\link{summary.model}}, \code{\link{predict.model}}
}
