% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{print.selection}
\alias{print.selection}
\title{Print a feature selection result}
\usage{
\method{print}{selection}(x, ...)
}
\arguments{
\item{x}{The result (object of class \code{\link{selection-class}}, created by
\code{\link{selectfeatures}}).}

\item{...}{Other parameters.}
}
\description{
Prints which features were selected, by which algorithm and criteria, instead of dumping the
underlying list.
}
\examples{
require (datasets)
data (iris)
selectfeatures (iris [, -5], iris [, 5], algorithm = "forward", multieval = "cfs")
}
\seealso{
\code{\link{selection-class}}, \code{\link{selectfeatures}}
}
