% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.som}
\alias{print.som}
\title{Print a self-organising map}
\usage{
\method{print}{som}(x, ...)
}
\arguments{
\item{x}{The map (object of class \code{\link{som-class}}, created by \code{\link{SOM}}).}

\item{...}{Other parameters.}
}
\description{
Prints the size of the map, how many of its units are actually used, and the dataset it was
fitted on.
}
\examples{
\donttest{
require (datasets)
data (iris)
SOM (iris [, -5], 4, 4)
}
}
\seealso{
\code{\link{som-class}}, \code{\link{SOM}}
}
