% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.spectral}
\alias{print.spectral}
\title{Print a spectral clustering}
\usage{
\method{print}{spectral}(x, ...)
}
\arguments{
\item{x}{The clustering (object of class \code{\link{spectral-class}}, created by
\code{\link{SPECTRAL}}).}

\item{...}{Other parameters.}
}
\description{
Prints the number of clusters found, their sizes and the dimension of the spectral
projection.
}
\examples{
\donttest{
require (datasets)
data (iris)
SPECTRAL (iris [, -5], 3)
}
}
\seealso{
\code{\link{spectral-class}}, \code{\link{SPECTRAL}}
}
