% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{pseudoF}
\alias{pseudoF}
\title{Pseudo-F}
\usage{
pseudoF(clustering)
}
\arguments{
\item{clustering}{The clustering result (obtained by the function \code{\link[stats]{kmeans}}).}
}
\value{
The pseudo-F of the clustering result.
}
\description{
Compute the pseudo-F of a clustering result obtained by the \emph{K}-means method.
}
\examples{
require (datasets)
data (iris)
km = KMEANS (iris [, -5], k = 3)
pseudoF (km)
}
\seealso{
\code{\link{kmeans.getk}}, \code{\link{KMEANS}}, \code{\link[stats]{kmeans}}
}
