% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{query.words}
\alias{query.words}
\title{Word query}
\usage{
query.words(wordvectors, origin, sub = NULL, add = NULL, nres = 5, lang = "en")
}
\arguments{
\item{wordvectors}{The vectorized words}

\item{origin}{The query (character).}

\item{sub}{Words to be substrated to the origin.}

\item{add}{Words to be Added to the origin.}

\item{nres}{The number of results.}

\item{lang}{The language of the words (NULL if no stemming).}
}
\value{
The Words the most similar to the query.
}
\description{
Search for words similar to the query.
}
\examples{
\donttest{
# 'capitals' is small, so the word vectors are coarse and 'ndim' is reduced
# accordingly; phrase detection needs a much larger corpus.
data (capitals)
words = vectorize.words (capitals, mincount = 2, ndim = 10, maxiter = 5)
query.words (words, origin = "paris", sub = "france", add = "germany")
query.words (words, origin = "berlin", sub = "germany", add = "france")
}
}
\seealso{
\code{\link{vectorize.words}}, \code{\link[text2vec]{sim2}}
}
