% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{reg1}
\alias{reg1}
\alias{reg1.train}
\alias{reg1.test}
\title{reg1 dataset}
\format{
50 instances and 3 variables. \code{X}, a numeric, \code{K}, a factor, and \code{Y}, a numeric (the target variable).
}
\usage{
reg1
reg1.train
reg1.test
}
\description{
Artificial dataset for simple regression tasks.
}
\author{
Alexandre Blansché \email{alexandre.blansche@univ-lorraine.fr}
}
