% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{reg2}
\alias{reg2}
\alias{reg2.train}
\alias{reg2.test}
\title{reg2 dataset}
\format{50 instances and 2 variables. \code{X} and \code{Y} (the target variable) are both numeric variables.}
\usage{
reg2
reg2.train
reg2.test
}
\description{
Artificial dataset for simple regression tasks.
}
\author{
Alexandre Blansché \email{alexandre.blansche@univ-lorraine.fr}
}
