% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rotation}
\alias{rotation}
\title{Rotation}
\usage{
rotation(d, angle, axis = 1:2, range = 2 * pi)
}
\arguments{
\item{d}{The dataset.}

\item{angle}{The angle of the rotation.}

\item{axis}{The axis.}

\item{range}{The range of the angle (360, 2*pi, 100, ...)}
}
\description{
Rotation on two variables of a numeric dataset
}
\examples{
d = data.parabol ()
d [, -3] = rotation (d [, -3], 45, range = 360)
plotdata (d [, -3], d [, 3])
}
