% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{runningtime}
\alias{runningtime}
\title{Running time}
\usage{
runningtime(FUN, ...)
}
\arguments{
\item{FUN}{The function to be evaluated.}

\item{...}{The parameters to be passes to function \code{FUN}.}
}
\value{
The running time of function \code{FUN}.
}
\description{
Return the running time of a function
}
\examples{
sqrt (x = 1:100)
runningtime (sqrt, x = 1:100)
}
\seealso{
\code{\link[base]{difftime}}
}
