% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{snore}
\alias{snore}
\title{Snore dataset}
\format{
The dataset has 100 instances described by 7 variables.
The variables are as follows:
\describe{
\item{\code{Age}}{In years.}
\item{\code{Weights}}{In kg.}
\item{\code{Height}}{In cm.}
\item{\code{Alcool}}{Number of glass of alcool per day.}
\item{\code{Sex}}{M for male or F for female.}
\item{\code{Snore}}{Snoring diagnosis (Y or N).}
\item{\code{Tobacco}}{Y or N.}
}
}
\source{
Originally published by G. Hunault, Departement Informatique, Universite d'Angers,
as the "RONFLE" file of his statistics dataset collection. That collection has changed
address twice and its current host was unreachable when this version was prepared, so the
citation is to an archived copy:
\url{https://web.archive.org/web/20250319122109/https://gilles-hunault.leria-info.univ-angers.fr/Datasets/datasets.htm}.
}
\usage{
snore
}
\description{
This dataset has been used in a study on snoring in Angers hospital.
}
