% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{spectral-class}
\alias{spectral-class}
\title{Spectral clustering model}
\description{
This class contains the model obtained by Spectral clustering.
}
\details{
Objects of this class are plain lists with the following components:
\describe{
  \item{\code{cluster}}{A \code{vector} of integer indicating the cluster to which each observation is allocated.}
  \item{\code{proj}}{The projection of the dataset in the spectral space.}
  \item{\code{centers}}{The cluster centers (on the spectral space).}
  \item{\code{data}}{The dataset that has been used to fit the model (as a \code{matrix}).}
}
}
\seealso{
\code{\link{SPECTRAL}}, \code{\link{predict.spectral}}
}
