% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{summary.model}
\alias{summary.model}
\title{Summary of a classification or regression model}
\usage{
\method{summary}{model}(object, ...)
}
\arguments{
\item{object}{The model (object of class \code{\link{model-class}}).}

\item{...}{Other parameters, passed to the summary of the wrapped model.}
}
\description{
Prints the short description of \code{\link{print.model}}, followed by the summary of the
wrapped model itself.
}
\examples{
require (datasets)
data (iris)
summary (NB (iris [, -5], iris [, 5]))
}
\seealso{
\code{\link{model-class}}, \code{\link{print.model}}
}
