% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SVD}
\alias{SVD}
\title{Singular Value Decomposition}
\usage{
SVD(x)
}
\arguments{
\item{x}{A numeric dataset (data.frame or matrix).}
}
\description{
Return the SVD decomposition.
}
\examples{
require (datasets)
data (iris)
SVD (iris [, -5])
}
\seealso{
code{\link[base]{svd}}
}
