% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Temperature dataset}
\format{The dataset has 35 instances described by 17 variables.
Average temperature of the 12 month. Mean and amplitude of the temperature. Latitude and longitude of the city. Localisation in Europe.}
\usage{
temperature
}
\description{
The data contains temperature measurement and geographic coordinates of 35 european cities.
}
