% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{TEXTMINING}
\alias{TEXTMINING}
\title{Text mining}
\usage{
TEXTMINING(corpus, miningmethod, vector = c("docs", "words"), ...)
}
\arguments{
\item{corpus}{The corpus.}

\item{miningmethod}{The data mining method.}

\item{vector}{Indicates the type of vectorization, documents (TF-IDF) or words (GloVe).}

\item{...}{Parameters passed to the vectorisation and to the data mining method.}
}
\value{
The result of the data mining method.
}
\description{
Apply data mining function on vectorized text
}
\examples{
\donttest{
require (text2vec)
# A small subset of movie_review is used here so this example runs quickly.
data ("movie_review")
d = movie_review [1:300, 2:3]
d [, 1] = factor (d [, 1])
d = splitdata (d, 1)
model = TEXTMINING (d$train.x, NB, labels = d$train.y, mincount = 10)
pred = predict (model, d$test.x)
evaluation (pred, d$test.y)
data (capitals)
clusters = TEXTMINING (capitals, HCA, vector = "words", k = 5, mincount = 2, ndim = 10, maxiter = 5)
plotclus (clusters$res, capitals, type = "tree", labels = TRUE)
}
}
\seealso{
\code{\link{predict.textmining}}, \code{\link{textmining-class}}, \code{\link{vectorize.docs}}, \code{\link{vectorize.words}}
}
