% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{exportgraphics.off}
\alias{exportgraphics.off}
\alias{exportgraphics.on}
\alias{toggleexport}
\alias{toggleexport.off}
\alias{toggleexport.on}
\title{Toggle graphic exports}
\usage{
exportgraphics.off()

exportgraphics.on()

toggleexport(export = NULL)

toggleexport.off()

toggleexport.on()
}
\arguments{
\item{export}{If \code{TRUE}, exports are activated, if \code{FALSE}, exports are deactivated. If \code{null}, switches on and off.}
}
\description{
Toggle graphic exports on and off
}
\examples{
\dontrun{
data (iris)
toggleexport (FALSE)
exportgraphics ("export.pdf")
plotdata (iris [, -5], iris [, 5])
closegraphics()
toggleexport (TRUE)
exportgraphics ("export.pdf")
plotdata (iris [, -5], iris [, 5])
closegraphics()
}
}
\seealso{
\code{\link{closegraphics}}, \code{\link{exportgraphics}}
}
