% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{vectorizer-class}
\alias{vectorizer-class}
\title{Document vectorization object}
\description{
This class contains a vectorization model for textual documents.
}
\details{
Objects of this class are plain lists with the following components:
\describe{
  \item{\code{vectorizer}}{The vectorizer.}
  \item{\code{transform}}{The transformation to be applied after vectorization (normalization, TF-IDF).}
  \item{\code{phrases}}{The phrase detection method.}
  \item{\code{tfidf}}{The TF-IDF transformation.}
  \item{\code{lsa}}{The LSA transformation.}
  \item{\code{tokens}}{The token from the original document.}
}
}
\seealso{
\code{\link{vectorize.docs}}, \code{\link{query.docs}}
}
