% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{zoo}
\alias{zoo}
\title{Zoo dataset}
\format{
The dataset has 101 instances described by 17 qualitative variables.
}
\source{
\url{https://archive.ics.uci.edu/dataset/111/zoo}
}
\usage{
zoo
}
\description{
Animal description based on various features.
}
