% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Alcohol dataset}
\format{The dataset has 27 instances described by 4 variables.
The variables are the average amount of alcool of different types per year par inhabitent.}
\source{
\url{https://www.who.int/}
}
\usage{
alcohol
}
\description{
This dataset has been extracted from the WHO database and depict the alcool habits in the 27 european contries (in 2010).
}
