% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\docType{class}
\name{apriori-class}
\alias{apriori-class}
\title{APRIORI classification model}
\description{
This class contains the classification model obtained by the APRIORI association rules method.
}
\section{Slots}{

\describe{
\item{\code{rules}}{The set of rules obtained by APRIORI.}

\item{\code{transactions}}{The training set as a \code{transaction} object.}

\item{\code{train}}{The training set (description). A \code{matrix} or \code{data.frame}.}

\item{\code{labels}}{Class labels of the training set. Either a \code{factor} or an integer \code{vector}.}

\item{\code{supp}}{The minimal support of an item set (numeric value).}

\item{\code{conf}}{The minimal confidence of an item set (numeric value).}
}}

\seealso{
\code{\link{APRIORI}}, \code{\link{predict.apriori}}, \code{\link{print.apriori}},
\code{\link{summary.apriori}}, \code{\link[arules]{apriori}}
}
