% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{autompg}
\alias{autompg}
\title{Auto MPG dataset}
\format{
The dataset has 392 instances described by 8 variables.
The seven first variables are numeric variables. The last variable is qualitative (car origin).
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/auto+mpg}
}
\usage{
autompg
}
\description{
This dataset was taken from the StatLib library which is maintained at Carnegie Mellon University.
The dataset was used in the 1983 American Statistical Association Exposition.
}
