% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{class}
\name{boosting-class}
\alias{boosting-class}
\title{Boosting methods model}
\description{
This class contains the classification model obtained by the CDA method.
}
\section{Slots}{

\describe{
\item{\code{models}}{List of models.}

\item{\code{x}}{The learning set.}

\item{\code{y}}{The target values.}
}}

\seealso{
\code{\link{ADABOOST}}, \code{\link{BAGGING}}, \code{\link{predict.boosting}}
}
