% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{class}
\name{cda-class}
\alias{cda-class}
\title{Canonical Disciminant Analysis model}
\description{
This class contains the classification model obtained by the CDA method.
}
\section{Slots}{

\describe{
\item{\code{proj}}{The projection of the dataset into the canonical base. A \code{data.frame}.}

\item{\code{transform}}{The transformation matrix between. A \code{matrix}.}

\item{\code{centers}}{Coordinates of the class centers. A \code{matrix}.}

\item{\code{within}}{The intr-class covarianc matrix. A \code{matrix}.}

\item{\code{eig}}{The eigen-values. A \code{matrix}.}

\item{\code{dim}}{The number of dimensions of the canonical base (numeric value).}

\item{\code{nb.classes}}{The number of clusters (numeric value).}

\item{\code{train}}{The training set (description). A \code{data.frame}.}

\item{\code{labels}}{Class labels of the training set. Either a \code{factor} or an integer \code{vector}.}

\item{\code{model}}{The prediction model.}
}}

\seealso{
\code{\link{CDA}}, \code{\link{plot.cda}}, \code{\link{predict.cda}}
}
