% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{CDA}
\alias{CDA}
\title{Classification using Canonical Discriminant Analysis}
\usage{
CDA(train, labels, tune = FALSE, ...)
}
\arguments{
\item{train}{The training set (description), as a \code{data.frame}.}

\item{labels}{Class labels of the training set (\code{vector} or \code{factor}).}

\item{tune}{If true, the function returns paramters instead of a classification model.}

\item{...}{Other parameters.}
}
\value{
The classification model, as an object of class \code{glmnet}.
}
\description{
This function builds a classification model using Canonical Discriminant Analysis.
}
\examples{
require (datasets)
data (iris)
CDA (iris [, -5], iris [, 5])
}
\seealso{
\code{\link{plot.cda}}, \code{\link{predict.cda}}, \code{\link{cda-class}}
}
