% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{closegraphics}
\alias{closegraphics}
\title{Close a graphics device}
\usage{
closegraphics()
}
\description{
Close the graphics device driver
}
\examples{
\dontrun{
data (iris)
exportgraphics ("export.pdf")
plotdata (iris [, -5], iris [, 5])
closegraphics()
}
}
\seealso{
\code{\link{exportgraphics}}, \code{\link{toggleexport}}, \code{\link[grDevices]{dev.off}}
}
