% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{compare.accuracy}
\alias{compare.accuracy}
\title{Comparison of two sets of clusters, using accuracy}
\usage{
compare.accuracy(clus, gt, comp = c("max", "pairwise", "cluster"))
}
\arguments{
\item{clus}{The extracted clusters.}

\item{gt}{The real clusters.}

\item{comp}{Indicates whether a "max" or a "pairwise" evaluation should be used, or the evaluation for each individual "cluster".}
}
\value{
A numeric value indicating how much the two sets of clusters are similar.
}
\description{
Comparison of two sets of clusters, using accuracy
}
\examples{
require (datasets)
data (iris)
km = KMEANS (iris [, -5], k = 3)
compare.accuracy (km$cluster, iris [, 5])
}
\seealso{
\code{\link{compare.jaccard}}, \code{\link{compare.kappa}}, \code{\link{compare}}
}
