% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{cookplot}
\alias{cookplot}
\title{Plot the Cook's distance of a linear regression model}
\usage{
cookplot(model, index = NULL)
}
\arguments{
\item{model}{The model to be plotted.}

\item{index}{The index of the variable used for for the x-axis.}
}
\description{
Plot the Cook's distance of a linear regression model.
}
\examples{
require (datasets)
data (trees)
model = LINREG (trees [, -3], trees [, 3])
cookplot (model)
}
