% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{data.diag}
\alias{data.diag}
\alias{data.square}
\title{Square dataset}
\usage{
data.square(
  n = 200,
  min = 0,
  max = 1,
  f = function(x) x,
  levels = NULL,
  graph = TRUE,
  seed = NULL
)
}
\arguments{
\item{n}{Number of observations in the dataset.}

\item{min}{Minimum value on each variables.}

\item{max}{Maximum value on each variables.}

\item{f}{The fucntion that separate the classes.}

\item{levels}{Name of each class.}

\item{graph}{A logical indicating whether or not a graphic should be plotted.}

\item{seed}{A specified seed for random number generation.}
}
\value{
A randomly generated dataset.
}
\description{
Generate a random dataset shaped like a square divided by a custom function
}
\examples{
data.square ()
}
\seealso{
\code{\link{data.parabol}}, \code{\link{data.target1}}, \code{\link{data.target2}}, \code{\link{data.twomoons}}, \code{\link{data.xor}}
}
