% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{data.target1}
\alias{data.target1}
\title{Target1 dataset}
\usage{
data.target1(r = 1:3, n = 200, sigma = 0.1, levels = NULL,
  graph = TRUE, seed = NULL)
}
\arguments{
\item{r}{Radius of each class.}

\item{n}{Number of observations in each class.}

\item{sigma}{Variance in each class.}

\item{levels}{Name of each class.}

\item{graph}{A logical indicating whether or not a graphic should be plotted.}

\item{seed}{A specified seed for random number generation.}
}
\value{
A randomly generated dataset.
}
\description{
Generate a random dataset shaped like a target.
}
\examples{
data.target1 ()
}
\seealso{
\code{\link{data.parabol}}, \code{\link{data.target2}}, \code{\link{data.twomoons}}
}
