% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{data.target2}
\alias{data.target2}
\title{Target2 dataset}
\usage{
data.target2(
  minr = c(0, 2),
  maxr = minr + 1,
  initn = 1000,
  levels = NULL,
  graph = TRUE,
  seed = NULL
)
}
\arguments{
\item{minr}{Minimum radius of each class.}

\item{maxr}{Maximum radius of each class.}

\item{initn}{Number of observations at the beginning of the generation process.}

\item{levels}{Name of each class.}

\item{graph}{A logical indicating whether or not a graphic should be plotted.}

\item{seed}{A specified seed for random number generation.}
}
\value{
A randomly generated dataset.
}
\description{
Generate a random dataset shaped like a target.
}
\examples{
data.target2 ()
}
\seealso{
\code{\link{data.diag}}, \code{\link{data.parabol}}, \code{\link{data.target1}}, \code{\link{data.twomoons}}, \code{\link{data.xor}}
}
