% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{data.twomoons}
\alias{data.twomoons}
\title{Two moons dataset}
\usage{
data.twomoons(r = 1, n = 200, sigma = 0.1, levels = NULL,
  graph = TRUE, seed = NULL)
}
\arguments{
\item{r}{Radius of each class.}

\item{n}{Number of observations in each class.}

\item{sigma}{Variance in each class.}

\item{levels}{Name of each class.}

\item{graph}{A logical indicating whether or not a graphic should be plotted.}

\item{seed}{A specified seed for random number generation.}
}
\value{
A randomly generated dataset.
}
\description{
Generate a random dataset shaped like two moons.
}
\seealso{
\code{\link{data.parabol}}, \code{\link{data.target1}}, \code{\link{data.target2}}
data.twomoons ()
}
