% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{dbs-class}
\alias{dbs-class}
\title{DBSCAN model}
\description{
This class contains the model obtained by the DBSCAN method.
}
\section{Slots}{

\describe{
\item{\code{cluster}}{A vector of integers indicating the cluster to which each point is allocated.}

\item{\code{eps}}{Reachability distance (parameter).}

\item{\code{MinPts}}{Reachability minimum no. of points (parameter).}

\item{\code{isseed}}{A logical vector indicating whether a point is a seed (not border, not noise).}

\item{\code{data}}{The dataset that has been used to fit the map (as a \code{matrix}).}
}}

\seealso{
\code{\link{DBSCAN}}
}
