% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{exportgraphics}
\alias{exportgraphics}
\title{Open a graphics device}
\usage{
exportgraphics(file, type = tail(strsplit(file, split = "\\\\.")[[1]], 1), ...)
}
\arguments{
\item{file}{A character string giving the name of the file.}

\item{type}{The type of graphics device.}

\item{...}{Other parameters.}
}
\description{
Starts the graphics device driver
}
\examples{
\dontrun{
data (iris)
exportgraphics ("export.pdf")
plotdata (iris [, -5], iris [, 5])
closegraphics()
}
}
\seealso{
\code{\link{closegraphics}}, \code{\link{toggleexport}}, \code{\link[grDevices]{Devices}}
}
