% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\docType{class}
\name{factorial-class}
\alias{factorial-class}
\title{Factorial analysis results}
\description{
This class contains the classification model obtained by the CDA method.
}
\seealso{
\code{\link{CA}}, \code{\link{MCA}}, \code{\link{PCA}}, \code{\link{plot.factorial}}
}
