% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{general.rules}
\alias{general.rules}
\title{Remove redundancy in a set of rules}
\usage{
general.rules(r)
}
\arguments{
\item{r}{A set of rules.}
}
\value{
A set of rules, without redundancy.
}
\description{
This function remove every redundant rules, keeping only the most general ones.
}
\examples{
require ("arules")
data ("Adult")
r = apriori (Adult)
inspect (general.rules (r))
}
\seealso{
\code{\link[arules]{apriori}}
}
