% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{HCA}
\alias{HCA}
\title{Hierarchical Cluster Analysis method}
\usage{
HCA(d, method = c("ward", "single"), k = NULL)
}
\arguments{
\item{d}{The dataset (\code{matrix} or \code{data.frame}).}

\item{method}{Character string defining the clustering method.}

\item{k}{The number of cluster.}
}
\value{
The cluster hierarchy (\code{hca} object).
}
\description{
Run the HCA method for clustering.
}
\examples{
require (datasets)
data (iris)
HCA (iris [, -5], method = "ward", k = 3)
}
\seealso{
\code{\link[cluster]{agnes}}
}
