% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ionosphere}
\alias{ionosphere}
\title{Ionosphere dataset}
\format{The dataset has 351 instances described by 34. The last variable is the class.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/ionosphere}
}
\usage{
ionosphere
}
\description{
This is a dataset from the UCI repository.
This radar data was collected by a system in Goose Bay, Labrador. This system consists of a phased array of 16 high-frequency antennas with a total transmitted power on the order of 6.4 kilowatts. See the paper for more details. The targets were free electrons in the ionosphere. "Good" radar returns are those showing evidence of some type of structure in the ionosphere. "Bad" returns are those that do not; their signals pass through the ionosphere.
Received signals were processed using an autocorrelation function whose arguments are the time of a pulse and the pulse number. There were 17 pulse numbers for the Goose Bay system. Instances in this databse are described by 2 attributes per pulse number, corresponding to the complex values returned by the function resulting from the complex electromagnetic signal.
One attribute with constant value has been removed.
}
