% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{KERREG}
\alias{KERREG}
\title{Kernel Regression}
\usage{
KERREG(x, y, bandwidth = 1, tune = FALSE, ...)
}
\arguments{
\item{x}{Predictor \code{matrix}.}

\item{y}{Response \code{vector}.}

\item{bandwidth}{The bandwidth parameter.}

\item{tune}{If true, the function returns paramters instead of a classification model.}

\item{...}{Other parameters.}
}
\value{
The classification model, as an object of class \code{\link{model-class}}.
}
\description{
This function builds a kernel regression model.
}
\examples{
require (datasets)
data (trees)
KERREG (trees [, -3], trees [, 3])
}
\seealso{
\code{\link[ibr]{npregress}}
}
