% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{KMEANS}
\alias{KMEANS}
\title{K-means method}
\usage{
KMEANS(d, k = 9, criterion = c("none", "pseudo-F"), graph = FALSE,
  nstart = 10, ...)
}
\arguments{
\item{d}{The dataset (\code{matrix} or \code{data.frame}).}

\item{k}{The number of cluster.}

\item{criterion}{The criterion for cluster number selection. If \code{none}, \code{k} is used, if not the number of cluster is selected between 2 and \code{k}.}

\item{graph}{A logical indicating whether or not a graphic should be plotted (cluster number selection).}

\item{nstart}{Define how many random sets should be chosen.}

\item{...}{Other parameters.}
}
\value{
The clustering (\code{kmeans} object).
}
\description{
Run K-means for clustering.
}
\examples{
require (datasets)
data (iris)
KMEANS (iris [, -5], k = 3)
KMEANS (iris [, -5], criterion = "pseudo-F") # With automatic detection of the nmber of clusters
}
\seealso{
\code{\link[stats]{kmeans}}, \code{\link{predict.kmeans}}
}
