% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{class}
\name{knn-class}
\alias{knn-class}
\title{K Nearest Neighbours model}
\description{
This class contains the classification model obtained by the k-NN method.
}
\section{Slots}{

\describe{
\item{\code{train}}{The training set (description). A \code{data.frame}.}

\item{\code{labels}}{Class labels of the training set. Either a \code{factor} or an integer \code{vector}.}

\item{\code{k}}{The \code{k} parameter.}
}}

\seealso{
\code{\link{KNN}}, \code{\link{predict.knn}}
}
