% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{KNN}
\alias{KNN}
\title{Classification using k-NN}
\usage{
KNN(train, labels, k = 1:10, tune = FALSE, ...)
}
\arguments{
\item{train}{The training set (description), as a \code{data.frame}.}

\item{labels}{Class labels of the training set (\code{vector} or \code{factor}).}

\item{k}{The k parameter.}

\item{tune}{If true, the function returns paramters instead of a classification model.}

\item{...}{Other parameters.}
}
\value{
The classification model.
}
\description{
This function builds a classification model using Logistic Regression.
}
\examples{
require (datasets)
data (iris)
KNN (iris [, -5], iris [, 5])
}
\seealso{
\code{\link[class]{knn}}
}
