% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{class}
\name{model-class}
\alias{model-class}
\title{Generic classification or regression model}
\description{
This is a wrapper class containing the classification model obtained by any classification or regression method.
}
\section{Slots}{

\describe{
\item{\code{model}}{The wrapped model.}

\item{\code{method}}{The name of the method.}
}}

\seealso{
\code{\link{predict.model}}, \code{\link[stats]{predict}}
}
