% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{movies}
\alias{movies}
\title{Movies dataset}
\format{
A set of 49 movies, rated by 55 users.
}
\source{
\url{https://grouplens.org/datasets/movielens/}
}
\usage{
movies
}
\description{
Extract from the movie lens dataset. Missing values have been imputed.
}
