% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{NMF}
\alias{NMF}
\title{Non-negative Matrix Factorization}
\usage{
NMF(x, rank = 2, nstart = 10, ...)
}
\arguments{
\item{x}{A numeric dataset (data.frame or matrix).}

\item{rank}{Specification of the factorization rank.}

\item{nstart}{How many random sets should be chosen?}

\item{...}{Other parameters.}
}
\description{
Return the NMF decomposition.
}
\examples{
require (datasets)
data (iris)
NMF (iris [, -5], nstart = 1)
}
\seealso{
\code{\link[NMF]{nmf}}
}
