% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{Ozone dataset}
\format{Each instance is described by the maximum level of ozone measured during the day.
Temperature, clouds, and wind are also recorded.}
\source{
\url{https://r-stat-sc-donnees.github.io/ozone.txt}
}
\usage{
ozone
}
\description{
This dataset constains measurements on ozone level.
}
