% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{class}
\name{params-class}
\alias{params-class}
\title{Learning Parameters}
\description{
This class contains main parameters for various learning methods.
}
\section{Slots}{

\describe{
\item{\code{decay}}{The decay parameter.}

\item{\code{hidden}}{The number of hidden nodes.}

\item{\code{epsilon}}{The epsilon parameter.}

\item{\code{gamma}}{The gamma parameter.}

\item{\code{cost}}{The cost parameter.}
}}

\seealso{
\code{\link{MLP}}, \code{\link{MLPREG}}, \code{\link{SVM}}, \code{\link{SVR}}
}
