% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{plot.cda}
\alias{plot.cda}
\title{Plot function for cda-class}
\usage{
\method{plot}{cda}(x, newdata = NULL, axes = 1:2, ...)
}
\arguments{
\item{x}{The classification model (object of class \code{cda-class}).}

\item{newdata}{The test set (\code{matrix} or \code{data.frame}).}

\item{axes}{The canonical axes to be printed (numeric \code{vector}).}

\item{...}{Other parameters.}
}
\description{
Plot the learning set (and test set) on the canonical axes obtained by Canonical Discriminant Analysis (function \code{CDA}).
}
\examples{
require (datasets)
data (iris)
model = CDA (iris [, -5], iris [, 5])
plot (model)
}
\seealso{
\code{\link{CDA}}, \code{\link{predict.cda}}, \code{\link{cda-class}}
}
