% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{plot.factorial}
\alias{plot.factorial}
\title{Plot function for factorial-class}
\usage{
\method{plot}{factorial}(x, type = c("ind", "cor", "eig"), axes = c(1, 2), ...)
}
\arguments{
\item{x}{The PCA, CA or MCA result (object of class \code{factorial-class}).}

\item{type}{The graph to plot.}

\item{axes}{The factorial axes to be printed (numeric \code{vector}).}

\item{...}{Other parameters.}
}
\description{
Plot PCA, CA or MCA.
}
\examples{
require (datasets)
data (iris)
pca = PCA (iris, quali.sup = 5)
plot (pca)
plot (pca, type = "cor")
plot (pca, type = "eig")
}
\seealso{
\code{\link{CA}}, \code{\link{MCA}}, \code{\link{PCA}}, \code{\link[FactoMineR]{plot.CA}}, \code{\link[FactoMineR]{plot.MCA}}, \code{\link[FactoMineR]{plot.PCA}}, \code{\link{factorial-class}}
}
