% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{plot.selection}
\alias{plot.selection}
\title{Plot a feature selection}
\usage{
\method{plot}{selection}(x, horiz = TRUE, legendpos = "bottomright", ...)
}
\arguments{
\item{x}{The selection (object of class \code{\link{selection-class}}, created by
\code{\link{selectfeatures}}).}

\item{horiz}{Whether the bars are drawn horizontally, which leaves room for long variable
names.}

\item{legendpos}{Position of the legend.}

\item{...}{Other parameters, passed to \code{\link[graphics]{barplot}}.}
}
\description{
Draws the score every variable obtained, the ones that were kept apart from the ones that
were not. Only a selection made by ranking the variables (\code{algorithm = "ranking"}) can
be drawn this way: it is the only one that scores them one by one, where the other three
score whole subsets.
}
\examples{
\donttest{
require (datasets)
data (iris)
# How useful a random forest finds each variable, and the two it would keep
selection = selectfeatures (iris [, -5], iris [, 5], unieval = "randomforest", uninb = 2)
selection
plot (selection)
}
}
\seealso{
\code{\link{selectfeatures}}, \code{\link{selection-class}},
\code{\link{print.selection}}
}
