% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotdata}
\alias{plotdata}
\title{Advanced plot function}
\usage{
plotdata(d, k = NULL, type = c("pairs", "scatter", "boxplot", "pca",
  "cda", "svd", "nmf", "tsne", "som"), legendpos = "topleft",
  alpha = 200, asp = 1, labels = FALSE, ...)
}
\arguments{
\item{d}{A numeric dataset (data.frame or matrix).}

\item{k}{A categorial variable (vector or factor).}

\item{type}{The type of graphic to be plotted.}

\item{legendpos}{Position of the legend}

\item{alpha}{Color opacity (0-255).}

\item{asp}{Aspect ratio (default: 1).}

\item{labels}{Indicates whether or not labels (row names) should be showned on the (scatter) plot.}

\item{...}{Other parameters.}
}
\description{
Plot a dataset.
}
\examples{
require (datasets)
data (iris)
# Without classification
plotdata (iris [, -5]) # Défault (pairs)
# With classification
plotdata (iris [, -5], iris [, 5]) # Défault (pairs)
plotdata (iris [, -5], iris [, 5], type = "scatter") # Scatter plot (PCA axis)
plotdata (iris [, -5], iris [, 5], type = "boxplot") # Boxplot
plotdata (iris [, -5], iris [, 5], type = "pca") # Scatter plot (PCA axis)
plotdata (iris [, -5], iris [, 5], type = "cda") # Scatter plot (CDA axis)
plotdata (iris [, -5], iris [, 5], type = "svd") # Scatter plot (SVD axis)
plotdata (iris [, -5], iris [, 5], type = "som") # Kohonen map
# With only one variable
plotdata (iris [, 1], iris [, 5]) # Défault (data vs. index)
plotdata (iris [, 1], iris [, 5], type = "scatter") # Scatter plot (data vs. index)
plotdata (iris [, 1], iris [, 5], type = "boxplot") # Boxplot
# With two variables
plotdata (iris [, 3:4], iris [, 5]) # Défault (scatter plot)
plotdata (iris [, 3:4], iris [, 5], type = "scatter") # Scatter plot
}
