% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{POLYREG}
\alias{POLYREG}
\title{Polynomial Regression}
\usage{
POLYREG(x, y, degree = 2, tune = FALSE, ...)
}
\arguments{
\item{x}{Predictor \code{matrix}.}

\item{y}{Response \code{vector}.}

\item{degree}{The polynom degree.}

\item{tune}{If true, the function returns paramters instead of a classification model.}

\item{...}{Other parameters.}
}
\value{
The classification model, as an object of class \code{\link{model-class}}.
}
\description{
This function builds a polynomial regression model.
}
\examples{
require (datasets)
data (trees)
POLYREG (trees [, -3], trees [, 3])
}
\seealso{
\code{\link[mda]{polyreg}}
}
