% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{QDA}
\alias{QDA}
\title{Classification using Quadratic Discriminant Analysis}
\usage{
QDA(train, labels, tune = FALSE, ...)
}
\arguments{
\item{train}{The training set (description), as a \code{data.frame}.}

\item{labels}{Class labels of the training set (\code{vector} or \code{factor}).}

\item{tune}{If true, the function returns paramters instead of a classification model.}

\item{...}{Other parameters.}
}
\value{
The classification model.
}
\description{
This function builds a classification model using Quadratic Discriminant Analysis.
}
\examples{
require (datasets)
data (iris)
QDA (iris [, -5], iris [, 5])
}
\seealso{
\code{\link[MASS]{qda}}
}
