% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regplot}
\alias{regplot}
\title{Plot function for a regression model}
\usage{
regplot(model, x, y, margin = 0.1, ...)
}
\arguments{
\item{model}{The model to be plotted.}

\item{x}{The predictor \code{vector}.}

\item{y}{The response \code{vector}.}

\item{margin}{A margin parameter.}

\item{...}{Other graphical parameters}
}
\description{
Plot a regresion model on a 2-D plot. The predictor \code{x} should be one-dimensional.
}
\examples{
require (datasets)
data (cars)
model = POLYREG (cars [, -2], cars [, 2])
regplot (model, cars [, -2], cars [, 2])
}
