% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\docType{class}
\name{selection-class}
\alias{selection-class}
\title{Feature selection}
\description{
This class contains the result of feature selection algorithms.
}
\section{Slots}{

\describe{
\item{\code{selection}}{A vector of integers indicating the selected features.}

\item{\code{unieval}}{The evaluation of the features (univariate).}

\item{\code{multieval}}{The evaluation of the selected features (multivariate).}

\item{\code{algorithm}}{The algorithm used to select features.}

\item{\code{univariate}}{The evaluation criterion (univariate).}

\item{\code{nbfeatures}}{The number of features to be kept.}

\item{\code{threshold}}{The threshold to decide whether a feature is kept or not..}

\item{\code{multivariate}}{The evaluation criterion (multivariate).}

\item{\code{dataset}}{The dataset described by the selected features only.}

\item{\code{model}}{The classification model.}
}}

\seealso{
\code{\link{FEATURESELECTION}}, \code{\link{predict.selection}}, \code{\link{selectfeatures}}
}
