% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{som-class}
\alias{som-class}
\title{Self-Organizing Maps model}
\description{
This class contains the model obtained by the SOM method.
}
\section{Slots}{

\describe{
\item{\code{som}}{An object of class \code{kohonen} representing the fitted map.}

\item{\code{nodes}}{A \code{vector} of integer indicating the cluster to which each node is allocated.}

\item{\code{cluster}}{A \code{vector} of integer indicating the cluster to which each observation is allocated.}

\item{\code{data}}{The dataset that has been used to fit the map (as a \code{matrix}).}
}}

\seealso{
\code{\link{plot.som}}, \code{\link{SOM}}, \code{\link[kohonen]{som}}
}
