% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{spectral-class}
\alias{spectral-class}
\title{Spectral clustering model}
\description{
This class contains the model obtained by Spectral clustering.
}
\section{Slots}{

\describe{
\item{\code{cluster}}{A \code{vector} of integer indicating the cluster to which each observation is allocated.}

\item{\code{proj}}{The projection of the dataset in the spectral space.}

\item{\code{centers}}{The cluster centers (on the spectral space).}
}}

\seealso{
\code{\link{SPECTRAL}}
}
