% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{SPECTRAL}
\alias{SPECTRAL}
\title{Spectral clustering method}
\usage{
SPECTRAL(d, k, sigma = 1, graph = TRUE, ...)
}
\arguments{
\item{d}{The dataset (\code{matrix} or \code{data.frame}).}

\item{k}{The number of cluster.}

\item{sigma}{Width of the gaussian used to build the affinity matrix.}

\item{graph}{A logical indicating whether or not a graphic should be plotted (projection on the spectral space of the affinity matrix).}

\item{...}{Other parameters.}
}
\description{
Run a Spectral clustering algorithm.
}
\examples{
require (datasets)
data (iris)
SPECTRAL (iris [, -5], k = 3)
}
\seealso{
\code{\link{spectral-class}}
}
