% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{spine}
\alias{spine}
\alias{spine.train}
\alias{spine.test}
\title{Spine dataset}
\format{The dataset has 310 instances described by 8 variables.
Variables V1 to V6 are biomechanical attributes derived from the shape and orientation of the pelvis and lumbar spine.
The variable Classif2 is the classification into two classes \code{AB} and \code{NO}.
The variable Classif3 is the classification into 3 classes \code{DH}, \code{SL} and \code{NO}.
\code{spine.train} contains 217 instances and \code{spine.test} contains 93.}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/vertebral+column}
}
\usage{
spine
spine.train
spine.test
}
\description{
The data have been organized in two different but related classification tasks.
The first task consists in classifying patients as belonging to one out of three categories: Normal, Disk Hernia or Spondylolisthesis.
For the second task, the categories Disk Hernia and Spondylolisthesis were merged into a single category labelled as 'abnormal'.
Thus, the second task consists in classifying patients as belonging to one out of two categories: Normal or Abnormal.
}
