% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{STUMP}
\alias{STUMP}
\title{Classification using one-level decision tree}
\usage{
STUMP(train, labels, minsplit = 1, cp = NULL, randomvar = TRUE,
  tune = FALSE, ...)
}
\arguments{
\item{train}{The training set (description), as a \code{data.frame}.}

\item{labels}{Class labels of the training set (\code{vector} or \code{factor}).}

\item{minsplit}{The minimum leaf size during the learning.}

\item{cp}{The complexity parameter of the tree. Cross-validation is used to determine optimal cp if NULL.}

\item{randomvar}{If true, the model uses a random variable.}

\item{tune}{If true, the function returns paramters instead of a classification model.}

\item{...}{Other parameters.}
}
\value{
The classification model.
}
\description{
This function builds a classification model using CART with maxdepth = 1.
}
\examples{
require (datasets)
data (iris)
STUMP (iris [, -5], iris [, 5])
}
\seealso{
\code{\link{CART}}
}
