% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\docType{class}
\name{textmining-class}
\alias{textmining-class}
\title{Text mining object}
\description{
Object used for text mining.
}
\section{Slots}{

\describe{
\item{\code{vectorizer}}{The vectorizer.}

\item{\code{vectors}}{The vectorized dataset.}

\item{\code{res}}{The result of the text mining method.}
}}

\seealso{
\code{\link{TEXTMINING}}, \code{\link{vectorize.docs}}
}
