% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic dataset}
\format{The dataset has 2201 instances described by 4 variables.
The variables are as follows:
\describe{
\item{\code{Category}}{1st, 2nd, 3rd Class or Crew.}
\item{\code{Age}}{Adult or Child.}
\item{\code{Sex}}{Female or Male.}
\item{\code{Fate}}{Casualty or Survivor.}
}}
\source{
British Board of Trade (1990), Report on the Loss of the ‘Titanic’ (S.S.). British Board of Trade Inquiry Report (reprint). Gloucester, UK: Allan Sutton Publishing.
}
\usage{
titanic
}
\description{
This dataset from the British Board of Trade depict the fate of the passengers and crew during the RMS Titanic disaster.
}
\seealso{
code{\link[datasets]{Titanic}}
}
