% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{TSNE}
\alias{TSNE}
\title{t-distributed Stochastic Neighbor Embedding}
\usage{
TSNE(x, perplexity = 30, nstart = 10, ...)
}
\arguments{
\item{x}{A numeric dataset (data.frame or matrix).}

\item{perplexity}{Specification of the perplexity.}

\item{nstart}{How many random sets should be chosen?}

\item{...}{Other parameters.}
}
\description{
Return the t-SNE dimensionality reduction.
}
\examples{
require (datasets)
data (iris)
TSNE (iris [, -5])
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
