% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{wheat}
\alias{wheat}
\title{Wheat dataset}
\format{The dataset has 210 instances described by 8 variables:
area, perimeter, compactness, length, width, asymmetry coefficient, groove length and variery.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/seeds}
}
\usage{
wheat
}
\description{
The data contains kernels belonging to three different varieties of wheat: Kama, Rosa and Canadian, 70 elements each, randomly selected.
High quality visualization of the internal kernel structure was detected using a soft X-ray technique. The images were recorded on 13x18 cm X-ray KODAK plates.
Source : Institute of Agrophysics of the Polish Academy of Sciences in Lublin.
}
