% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine dataset}
\format{
There are 178 observations and 14 variables.
The first variable is the class label (\code{1}, \code{2}, \code{3}).
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/wine}
}
\usage{
wine
}
\description{
These data are the results of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars.
The analysis determined the quantities of 13 constituents found in each of the three types of wines.
}
